Method 2 Class dependency
A new scheme of knowledge encoding in a fuzzy
multilayer perceptron (MLP) using rough set-theoretic concepts
is described. Crude domain knowledge is extracted from the data
set in the form of rules. The syntax of these rules automatically
determines the appropriate number of hidden nodes while the
dependency factors are used in the initial weight encoding. The
network is then refined during training.